ipmf/requirements.txt
# Core Django
Django>=5.0,<5.2

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# CORS & Security
django-cors-headers==4.4.0

# Filtres
django-filter==24.3

# Import/Export
django-import-export==4.1.1

# Images
Pillow>=11.0.0

# Utils
django-extensions==3.2.3
python-decouple==3.8
psycopg2-binary==2.9.9
google-generativeai==0.8.3

# === PRODUCTION ===
# Serveur WSGI de production (remplace runserver)
gunicorn==21.2.0

# Servir les fichiers statiques sans Nginx
whitenoise==6.7.0

# Parse l'URL de la base de données Render (DATABASE_URL)
dj-database-url==2.2.0

# Stockage des fichiers média sur le cloud (photos, pièces jointes)
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
